requirements.txt
tweepy
PyYAML>=5.3
